Home Screen - Show Featured assessments onclick shows detail view.
