+++ lib/core/assesments/data/repository/assesment_repository.dart
@@ -1,4 +1,4 @@
-import 'package:exagram/feature/assessment_detail/page/assessment_detail_controller.dart';
+import 'package:exagram/feature/assessment_detail/assessment_detail_controller.dart';
 
 import '../../domain/entities/featured_assessment_card_dto.dart';
 import '../../domain/entities/free_tier_assessment.dart';
