+++ lib/core/assesments/domain/usecase/get_assessment_detail_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:exagram/core/assesments/data/repository/assesment_repository.dart';
-import 'package:exagram/feature/assessment_detail/page/assessment_detail_controller.dart';
+import 'package:exagram/feature/assessment_detail/assessment_detail_controller.dart';
 
 class GetAssessmentDetailUseCase {
   final AssessmentRepository assessmentRepository;
