--- lib/feature/counter/CounterController.dart
@@ -1,8 +1,0 @@
-import 'package:get/get.dart';
-import 'package:get/get_state_manager/get_state_manager.dart';
-
-class CounterController extends GetxController {
-  var count = 0.obs;
-
-  increment() => count++;
-}
